requirements.txt
future>=0.18.3
Pillow>=9.3.0